Add species/background newgame doll generator
This adds a helper function to generate dolls for character combos.
This is done in a rather ugly fashion by (partially) generating
the new game character. A small segment of the character setup
code in _setup_generic() that's relevant to initial equipment is
duplicated, with some minor changes; this code modifies the global `you`
variable, and fill_doll_equipment() is then called to generate the doll
corresponding to the character in `you`.

Regarding the modified character generation code, everything after
give_items_skills() is removed, since that's what determines a player's
starting equipment, and thus the doll appearance. An exception is
the item colour randomization. Separately, _unfocus_stats() was not
included, since I'm reasonably sure that this has no effect on initial
item generation, and I didn't want to make it non-static.